+++ srcs/ft_min.c
@@ -3,14 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_min.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: nmougino <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: nmougino <nmougino@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/03/21 19:01:27 by nmougino          #+#    #+#             */
-/*   Updated: 2016/03/21 19:39:42 by nmougino         ###   ########.fr       */
+/*   Updated: 2016/09/29 03:06:49 by nmougino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_min(int a, int b)
+#include "libft.h"
+
+intmax_t	ft_min(intmax_t a, intmax_t b)
 {
 	if (a <= b)
 		return (a);
